python cleanup
