+++ MANIFEST.in
@@ -1,0 +1,4 @@
+include CHANGELOG
+include LICENSE
+include README.md
+include mnet.conf
