+++ mnetsuite/util.py
